Moved docmap steps generation to `docmaps_steps.py` from `docmaps.py`
